Pre-commit cleanup
This moves enums to their own file
